StudentProjectsDetailController moved to new folder Student 	renamed:    app/Http/Controllers/api/StudentProjectsDetailController.php -> app/Http/Controllers/api/Student/StudentProjectsDetailController.php 	modified:   routes/api/v1.php 	modified:   tests/Feature/Student/StudentProjectsDetailControllerTest.php
