+++ app/Http/Controllers/api/Student/StudentProjectsDetailController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Controllers\api;
+namespace App\Http\Controllers\api\Student;
 
 use App\Http\Controllers\Controller;
 use App\Service\Student\StudentProjectsDetailService;
